+++ client/components/paragraphs/test/index.tsx
@@ -3,6 +3,7 @@
 /**
  * External dependencies
  */
+import * as React from 'react';
 import { render } from '@testing-library/react';
 
 /**
@@ -12,13 +13,17 @@ import Paragraphs from '../';
 
 describe( 'Paragraphs', () => {
 	test( 'renders set of strings correctly.', () => {
-		const strings = [ 'Paragraph 1', 'Paragraph 2', 'Paragraph 3' ];
+		const strings: string[] = [
+			'Paragraph 1',
+			'Paragraph 2',
+			'Paragraph 3',
+		];
 		const { container } = render( <Paragraphs>{ strings }</Paragraphs> );
 		expect( container ).toMatchSnapshot();
 	} );
 
 	test( 'renders empty set of strings correctly', () => {
-		const strings = [];
+		const strings: string[] = [];
 		const { container } = render( <Paragraphs>{ strings }</Paragraphs> );
 		expect( container ).toMatchSnapshot();
 	} );
